Update ex.rst
Giving an extra hint for the students to use the -h option in order to figure out how to use dftd3 and gcp. 
It might occur that an outdated gcp version is called upon typing (gcp) which does not yet have the def-TZVP basis set implemented. In these cases I also added the whole path to the right gcp version in the AK-bin.